+++ instrumentation/net/http/otelhttp/config.go
@@ -29,7 +29,6 @@ type config struct {
 	PublicEndpointFn  func(*http.Request) bool
 	ReadEvent         bool
 	WriteEvent        bool
-	WithoutRouteAttr  bool
 	Filters           []Filter
 	SpanNameFormatter func(string, *http.Request) string
 	ClientTrace       func(context.Context) *httptrace.ClientTrace
@@ -162,13 +161,6 @@ func WithMessageEvents(events ...Event) Option {
 				c.WriteEvent = true
 			}
 		}
-	})
-}
-
-// WithoutMetricRouteAttribute returns an [Option] that disables recording the HTTP route path as a metric attribute.
-func WithoutMetricRouteAttribute() Option {
-	return optionFunc(func(c *config) {
-		c.WithoutRouteAttr = true
 	})
 }